Cut the precision

+++ src/main/java/kubatech/tileentity/gregtech/multiblock/GT_MetaTileEntity_MegaIndustrialApiary.java
@@ -390,7 +390,7 @@ public class GT_MetaTileEntity_MegaIndustrialApiary
             builder.append(mStorage.get(i).queenStack.getDisplayName());
             builder.append(EnumChatFormatting.GREEN);
             mStorage.get(i).dropProgress.values().forEach(v -> builder.append(" ")
-                    .append(String.format("%.2f%%", v * 100d)));
+                    .append(String.format("%d%%", (int) (v * 100d))));
             info.add(builder.toString());
         }
